build: Don't ignore `main`
This is a dangerous name to ignore -- basically any Java/Scala code will
be in a directory like `src/main` which will be entirely ignored by this
pattern.

We should either ignore `/main` (if we really want to ignore it only at
the root level) or better yet change the configuration of `wren` to not
compile to `main`.