feat: introduce pipeline scheduler
This is part of #409.

Introduces `Pipeline` information to the physical plan. This indicates
which steps are part of a linear sequence, and should (ideally) be
executed together.

Also implements a pipeline "scheduler" to determine the pipeline for
each step, in a new `sparrow-backend` crate. As the physical plan is
built-up, the code should go in this "compiler backend" package, which
can own optimization and conversion of logical plans to physical plans.