feat: expression evaluation for new plans
This introduces the basic evaluator mechanisms for the new physical
plans. As part of this, some aspects of the physical expression
representation changed. Specifically, allowing a function call to
have zero or more literal arguments allowed every expression to be
"simple".

This also copied most of the existing evaluators into the new API.
This was done by copying to avoid overhead involved in refactoring.
The plan is to get the new runtime path based on these working (and
passing the same tests) and then delete the old path.

This doesn't copy over some of the scalar functions (comparisons) or
all of the time functions. Those can be ported over in the future once
the runtime is in place.

This also doesn't copy over any stateful operations (aggregations) yet
since aggregation may need to use a different approach to interface with
the physical plans.