feat: Introduce a Batch wrapper
Currently, we use a few different batches in different places during
execution. This doesn't yet replace any of those, however it introduces
a single batch representation that supports the multiple uses. This is
based on the batch used in a prototype that plumbs a single kind of
batch everywhere in execution.

Future PRs will update existing code and/or pull in prototype code using
the Batch, with the goal of replacing the different batch
representations.