feat: Use object_store to read prepare inputs
For CSV, this uses object store to copy the file to
local disk, since we need to infer the schema from the
entire file, and the CSV inference doesn't support a
streaming mode.

For Parquet, this uses streaming, async reads.

This is part of #465.

This also introduced `ObjectMetaExt` to make it possible to get a
hash of the etag.