wip - trying to implement Python UDF's
The problem here is going to be exposing the data. Polars uses Arrow2,
which may be compatible with Arrow, but not in an obvious way. Pyarrow
doesn't appear to have bindings for PyO3. There's a library exposing
datafusion to python, and that might be usable - I think DataFusion uses
the same arrow version we're using: https://github.com/apache/arrow-datafusion-python