feat: add prototype python query builder
This adds a new top-level crate `sparrow-py`. It is intentionally
outside of `crates` since it needs to be in a separate workspace to
properly compile the Python code.

This supports basic builders, implements most of the math operators
directly on expressions, and supports a `pipe` function similar to
Pandas.

In Python 3.11 the exceptions are printed with locations highlighted and
type information of each argument is printed.