feat: Execute Python queries
This removes the ability to inspect the prepared data from the
Python FFI. Instead, the Python FFI now has an `execute` method
allowing you to inspect the results of an expression, including
the bare table.

This changes the execute methods a bit to have a simpler API to
interface with.

This relies on pushing "in-memory" data into the `DataContext`,
which is very brittle and hacky. For example, it currently only
supports adding data once, and requires that the schemas line up.