bug: fix default schema in inference for map/list (#577)
Fixes the default schema for the `map` and `list` types in inference.
This isn't a good solution, as there's no guarantee that Arrow doesn't
change this behavior in the future. Ideally, we can figure out how to
plumb the user map type information through inference. But, it's
difficult because functions that produce `map` or `list` types need an
arbitrary naming/nullability that we define.

Closes #576