feat: Support aggregation functions in Python (#586)
Introduces Python classes for representing windows and flattens them
into arguments for the corresponding aggregation functions.

This also changes the AST to DFG logic so that it only tries to flatten
if the original AST is available. This does mean that ticks won't be
recreated, meaning they won't have the proper domain.

Both of these likely need to be extended to cover other aggregations.