feat: More work on Python API (#598)
- Introduce options for limits, and the method `show()` for displaying
the results of a query. Will use `ipython.display()` if available, or
print the Pandas dataframe to stdout otherwise.
- Added pydoc for collect
- Fixed some issues in the pytest for adding data (specifically around
how the `_time` column round-tripped through the JSON in the golden
test).
- Updated some of the modules to allow `import sparrow_py as kt` and
using `kt.sources.CsvSource`. The trick was to make sure the root module
(`sparrow_py` currently) re-exports the submodule (`sources`).