feat: add min parameter to collect (#607)
Adds the `min` parameter to collect, which allows a user to specify a
minimum list length before a non-null value is produced from
`collect()`. also reorders parameters to make the python builder arg
pattern a little cleaner.

Unfortunately, we still aren't able to use `collect` as a replacement
for `lag`, given the current behavior. `Lag` does not count `null` as a
valid value, while `collect` does. See the ignored unit test for an
example of differences.