feat: Update key hash inverses as new data arrives
Previously, we froze the key hash inverse based on the data that
was available when a query started (by computing it from the
merged data). Now, we compute it as data is added and maintain it.

Note: This does mean that if a `with_key` runs with the same grouping
as an existing table, any keys it adds will be added to the in-memory
key hash inverse.