feat: Cleanup and extend sources
This drops the `Source` from the name, since we use them as
`kt.sources.CsvString`.

This introduces `JsonlString` (for line-delimited JSON strings),
`Pandas` for pandas dataframes, `Parquet` for Parquet data, and `Pylist`
for lists of Python dicts.

The schema can be specified in the constructor. This allows creating
two sources using the same schema, by creating one and using it's
schema when creating the other. Additional data added to the source is
requested to conform to the original schema.