feat: Introduce crate for standalone signatures (#706)
This provides a standalone signature parser (without location tracking)
and instantiation. The intention is to provide the core type-checking
component from the compiler in a stand-alone component for
`sparrow-session` (and possibly checking logical/physical plans).

This doesn't use any fancy lexer/parser -- instead it just parses the
string directly. This also doesn't use names in for the function,
assuming that isn't necessary in the new direction. This also doesn't
track locations, etc.