feat: allow comparisons against literal timedeltas (#731)
This adds in a way to compare `timedeltas (durations)`. Arrow does not
currently support casting any duration types, so this hacks in a manual
math + cast to convert from less granular to more granular duration
types.

Note that you cannot perform operations between `timedeltas` and
`timestamps`. This is expected, as they are different concepts (one is
the time between two points, one is an absolute point in time)