bug: drains output channel when batching results  (#763)
Drains the output channel. Previously was causing queries to block on
send.

Puts the tokio runtime into an `Arc` to share it -- the `handle()` was
not working like I wanted since it was being dropped (a blocking method)
from an asynchronous context. Unsure why, but swapping to use `Arc`, as
is also recommended, works fine.

https://docs.rs/tokio/latest/tokio/runtime/struct.Runtime.html#sharing

Closes #775

---------

Co-authored-by: Ben Chambers <35960+bjchambers@users.noreply.github.com>