ref: Simplify expressions in physical plan (#782)
Change expressions to always produce the result of the expression. This
simplifies evaluation, without affecting the expressiviness:

1. We can now produce a primitive result (such as a boolean or integer)
if that is the type of the final expression.
2. We can still produce records by adding a `record` expression to the
end (which is the standard way of creating intermediate records).
3. We can still reference parts of a record input using `FieldRef`.

This eliminates the weird "some expressions are marked as output"
pattern. It also means that we can operate on `ArrayRef` generally and
use `DataType` as the type, instead of sometimes using `RecordBatch` and
`SchemaRef`.

Other smaller changes in this PR:
- Move the `Expressions` out of the `StepKind`. This allows the
`StepKind` to serve as an identifier of the step, regardless of the
expressions it computes, simplifying some of the de-duplication work in
the logical-to-physical compilation.
- Rename the `column` expression to `input`, since it accesses the input
data to the expression evaluation.
- Addition of a method to "intern" expression names, replacing a `&str`
with `&'static str` for expression operators available in the binary.