ref: move prepared files to ~/.cache dir (#781)
Moves the prepared files directory to `~/.cache/kaskada/`. 

The other alternative is to, for now, create a temporary directory that
cleans up after use. While the `~/.cache` is generally used for
application-specific files (which is what the prepared files are), we
aren't utilizing them and won't until we have state. But, if it's not a
large concern to accumulate prepared files in the cache, we might as
well start with it here in anticipation of using them.

Closes #779