ref: split up the expression crate
This moves the traits for defining an expression evaluator into
`sparrow-interfaces` and the code for executing a sequence of
expressions into `sparrow-expr-execution.`.

This allows flattening the `sparrow-expressions` crate, and only using
it as a test dependency in a wide variety of places. It is only needed
at execution time.

This is preparation for adding `sparrow-aggregations` to implement the
stateful aggregations.